ci: run import progress regression
